Fix for a small bug with popups.
Simplest reproduction:

    $('body').popup({
    });

This caused the error message:

    Uncaught TypeError: Cannot call method 'unbind' of undefined jq.ui.js:1717

This is because the default doneCallback did a "self.hide()" at line 1692, which is also set if "autoCloseDone" is set (see line 1752).

It works fine if you use:

    $('body').popup({
        doneCallback: function() {}
    });

My fix is to remove the self.hide call at line 1692. You might have a better fix, or I may be missing something!