Use jsdom rather than domino
domino doesn't support the full DOM API; in particular,
the "elements" property on form elements is missing.

jsdom does a better job of supporting the full DOM API, but
requires an older version of node than I was using. So I changed
the node engine requirements in package.json and .travis.yml
to ensure that tests are run with the correct node version.