$.getCssMatrix() should always return a matrix
If no element is supplied to $.getCssMatrix(), and either of
window.WebKitCSSMatrix or window.MSCSSMatrix are set, the
function returns a function rather than a matrix.

To ensure consistency and no surprises, change the method so
that it always returns a matrix object (fake or real).

In addition, previously, if there is no transform property available
in the computed style object, an exception would be thrown,
and caught to return a default matrix. Rather than using
exceptions to control flow, check for the existence of a transform
string before attempting to process it. If not available,
default to returning an "empty" matrix.