Add simple test for jsonp()
This tests the case where the jsonp request is made with
a callback name hard-coded into the returned script file.

The other use case for jsonp is where the script is requested
with a querystring like "http://host/script?callback=mycallback".
The server then returns a script with the callback name
from the querystring embedded, e.g. "mycallback({a: 1, b: 2});".

As this requires processing on the server-side (to extract
the querystring variable and customise the content of the
returned script file), there's no test for this case yet.