Update af.ui.js
Fix related with issue #837.

On first load of an Ajax Call, the href of the footer button will be = to the variable "hash" but without any hash character (#) at the beginning. Thus, the panel will correctly load with the results of an Ajax Call but the "pressed" class won't be added. The code tries to find an href having a # at the beginning (which is not the case on Ajax calls).

Last fix almost cleared all issues regarding footer buttons not being shown as pressed. Except for one case. If your app starts somewhere such as app.com/A then the footer button for "A" won't be shown as pressed in the first instance. Further clicks will work OK.

This fixes the issues for all cases.